requirements.txt
fastapi
uvicorn
python-dotenv
redis
rq
sqlalchemy
psycopg2-binary
jwt
